Enable stats by type instead of global and code enhancements
Stats can now be enabled and disabled by type instead of a global
setting. This way you can, for example, only disable the second and
third stat, but keep the rest. A lot of code enhancements have been made
as well, which will hopefully result in a better and smoother experience
with the plugin.