Add random join signs
These random join signs will join an arena which is: a) not in-game and
b) not full. Then it'll chose the arena with the most amount of players
starting with the first one in the list. The random join signs can be
created by typing '[buildinggame]' on line 1 and 'join' on line 2. The
third line should not equal the name of an already existing arena and
line 4 can be anything you want. There are also 4 setting sadded to the
messages.yml for the lines of the random join signs.