Fix stat to file saving and update to 1.12
The list which contains the statistics is now synchronized, so the async
saving process should work correrclty. The plugin has been fully updated
to 1.12 which resulted in some small changes. Mainly the fireworks
command being different again, so now there will be lots of people
complaining that those commands don't work. Hopefully that should be
over in a week.