Fix ItemBuilder issues and bump to version 5.0.0
Not overriding the equals and hashcode methods led to duplicate results
which meant that listeners that should be registered would be
unregistered. An implementation for both these methods solved the
problem. Also bumped to version 5.0.0 as this was the last part of the
update. Good luck breaking the plugin again!