Make players spawn in a safe spot while voting
Previously we teleported the players randomly in the plot which could
mean they got stuck into the build of players. Now an algorithm has been
implemented which looks for a safe spot to spawn the players on, meaning
they won't get stuck in blocks again. The only exception to this rule is
when there isn't a safe spot to spawn, then we'll just teleport the
players to a random spot like usual. However if this is the case the
players have either build something terrible, your plot is too small, or
your timer is too long.