5.2.1 (#85)
* Add PerWorldInventory compatibility fix

It was commented out, because the update wasn;t out yet. Now it is, so
we can introduce the fix again. Some version checking has been added as
well.

* Fix NPE with breaking join signs

The signs didn't check if the sign was actually a join sign causing
issues when checking the non-existent arena.

* Reload timers whenc reating an arena

Timers weren;t set correctly for newly created arenas.

* V 5.2.1