Switch to XML-based IF
The old gui system is replaced by this new adaption of IF with XML-file
based gui loading. The new system allows gui creation either by code or
XML-file or a combination of both. This allows us to create and change
guis much more rapidly without the need to continuously mangae
individual items. The implementation of IF here in combination with the
implementation of IF shown in an earlier stage combined with new
addition will be released as a standalone framework in the future (if
everything goes as planned). When this is released we will delete the
current implementation and switch over to the new one, but for now we
manage the implementation ourselves.