Add floor changers using Citizens
These can be placed and allow users to edit the floor by clicking on
them. Clicking with an item in your hand will use that item as the
floor; clicking with nothing in your hand will open a menu with blocks
to choose from.

When an NPC is placed inside a plot, the user that did so will
automatically be asked if they want this NPC be capable of changing the
floor. This functionality can be enabled at a later point for any NPC by
adding the buildinggame:floorchange trait to the NPC by using Citizens
commands.